Optionally always use the included libonig sources.
By default, the build process will check if libonig has already been
installed and if so, use that.  Otherwise, it will compile the copy
included with the gem and use that instead.

This is not always desirable, e.g. when compiling a redistributable
program that can't be shipped with the shared library.

With this change, if environment variable MRUBY_ONIGMO_USE_INTERNAL is
set and not empty, the build process will always use the included
library regardless of what else is available.

(This change also adds a Rakefile and build_config.rb that will build
mruby in the current directory and add this gem to it.  This is a
development aid.)